hopefully fixed CI by building qemu ourselves

+++ kernel/src/common/include/interrupt.h
@@ -92,7 +92,7 @@ enum SemihostingSWI {
 
 void SemihostingCall(enum SemihostingSWI mode);
 
-void SemihostingOSExit(int code) __attribute__ ((noreturn));;
+void SemihostingOSExit(uint8_t code) __attribute__ ((noreturn));;
 
 typedef enum {
     IRQ,		// (this is bit 0x8 on the CPSR)
